+++ frontend/src/components/ui/FileUploader.vue
@@ -59,12 +59,14 @@
       <input type="file" ref="fileInput" style="display: none" @change="handleFileChange" />
       <div class="dropzone-content">
         <img :src="UploadIllustration" alt="Upload Icon" class="upload-icon" />
-        <p class="drag-drop-text">Drag and drop or click to <span class="upload-link">upload file</span></p>
-        <p class="important-text">
-          <span class="important-label">Important:</span> The name of the document you upload will display on the client
-          portal. Please carefully check your file name before uploading to avoid
-          errors and duplicates.
-        </p>
+        <slot name="label">
+          <p class="drag-drop-text">Drag and drop or click to <span class="upload-link">upload file</span></p>
+          <p class="important-text">
+            <span class="important-label">Important:</span> The name of the document you upload will display on the client
+            portal. Please carefully check your file name before uploading to avoid
+            errors and duplicates.
+          </p>
+        </slot>
       </div>
     </div>
   </div>